Improve integration with MEAI typical usage pattern
In MEAI, the expectation is that you get chat clients for specific models from the same OpenAI client. There's no concept of a default model at that level.

This aligns the implementation with that approach.

In addition, we add an analyzer that ensures the user can still invoke `GetChatClient(model).AsIChatClient()` while getting the proper behavior which in our `GrokClient` case is that you should use the returned client from `GetChatClient(model)` directly without an additional wrapping done by the MEAI `AsIChatClient`, since we implement the `IChatClient` in the returned client directly.

This allows us to cache and use multiple chat clients dynamically matching the model in each request's ChatOptions.

This also fixes a bug where we were not properly setting the search mode to On instead of Auto when using Tools.