Add tool-focused extension to inspect calls and results
It's sometimes useful to be able to inspect (wether in tests or in production) the invocations that were performed by the model, including typed results.

The introduced `ToolJsonOptions` provide a mechanism to automatically inject a `$type` for result types so they can be later inspected by the `FindCall<TResult>` extension for `ChatResponse`.

To simplify this scenario, we also provide a `ToolFactory` that automatically sets things up for this scenario, additionally making the tool name default to the naming convention for tools, rather than the .NET method name.