Friendly error when non-matching SDK is used
Since we are building with latest&greatest .NET to get the extension class syntax, consumers need to use a matching SDK in order for that to work.

Provide a compile-time error instead of the inscrutable alternative (unsupported language feature or what-not).

This will turn off itself when we ship with a stable SDK.