Initial support for configurable AI agents
Following the pattern from configurable chat clients. Unlike chat clients, agents use the name property as the key (enforced by the underlying Agents.AI.Hosting library).

We rename the UseChatClients to AddChatClients since that's more consistent with the Add/Use pattern: the former is for registering stuff in the collection prior to service provider/app build, the latter to configure/start injected services on top of the added services.