Add support for setting reasoning effort and verbosity via config
We already support those properties via extensions in our Extensions.AI project, but agent was not setting them properly.

This is otherwise quite cumbersome to set (see https://github.com/microsoft/agent-framework/issues/1455).

We also turn on configuration binder source generation and review the emitted warnings for properties that wouldn't be deserialized (since they can't be easily mapped) and ignore the warnings/errors accordingly. We might want to offer custom converters for them in the future.