Add support for configurable and composable AI contexts
In order to make configured agents more dynamic and extensible, we introduce the `use` setting which is an array of named AI contexts to include as part of the agent execution.

These cannot be used at the same time with AIContextProvider (for now?) to simplify usage. You either pick exporting your own context provider (or factory), or you rely on more granular and reusable AIContext exported services.