Add support for instructions in markdown files with YAML front-matter
If instructions are long and the primary content that defines an agent, using TOML is still an inconvenience. For longer text, markdown is king.

We add support for markdown instructions by just relying on YAML front-matter to provide the section data to go along with it.

Example:

```
---
id: ai.agents.notes
description: Provides free-form memory
client: Grok
options:
  modelid: grok-4-fast
---
You organize and keep notes for the user
```