Make it possible to construct the raw completion request from gRPC
This is the pattern followed by the OpenAI implementation too. Allows exposing ALL the underlying options, albeit quite a bit less discoverable.

This should be sufficient fallback for any new (auto-updated) features in the upstream GrokClient generated from .proto files if we ever lag behind it on this library.

We make the test fully isolated from xAI calls to avoid spending tokens on a test that doesn't really require hitting the backend (unlike most of the others).